rename Tabs -> InterestTabs
