+++ src/components/InterestTabs.tsx
